python/web_service/requirements.txt
fastapi[standard]
jinja2
pyjwt